+++ x/foundation/msgs.go
@@ -25,7 +25,7 @@ func (m MsgFundTreasury) ValidateBasic() error {
 
 // GetSigners implements Msg.
 func (m MsgFundTreasury) GetSigners() []sdk.AccAddress {
-	signer, _ := sdk.AccAddressFromBech32(m.From)
+	signer := sdk.MustAccAddressFromBech32(m.From)
 	return []sdk.AccAddress{signer}
 }
 
@@ -50,7 +50,7 @@ func (m MsgWithdrawFromTreasury) ValidateBasic() error {
 
 // GetSigners implements Msg.
 func (m MsgWithdrawFromTreasury) GetSigners() []sdk.AccAddress {
-	signer, _ := sdk.AccAddressFromBech32(m.Operator)
+	signer := sdk.MustAccAddressFromBech32(m.Operator)
 	return []sdk.AccAddress{signer}
 }
 
@@ -75,7 +75,7 @@ func (m MsgUpdateMembers) ValidateBasic() error {
 
 // GetSigners implements Msg.
 func (m MsgUpdateMembers) GetSigners() []sdk.AccAddress {
-	signer, _ := sdk.AccAddressFromBech32(m.Operator)
+	signer := sdk.MustAccAddressFromBech32(m.Operator)
 	return []sdk.AccAddress{signer}
 }
 
@@ -99,7 +99,7 @@ func (m MsgUpdateDecisionPolicy) ValidateBasic() error {
 
 // GetSigners implements Msg.
 func (m MsgUpdateDecisionPolicy) GetSigners() []sdk.AccAddress {
-	signer, _ := sdk.AccAddressFromBech32(m.Operator)
+	signer := sdk.MustAccAddressFromBech32(m.Operator)
 	return []sdk.AccAddress{signer}
 }
 
@@ -179,7 +179,7 @@ func (m MsgSubmitProposal) UnpackInterfaces(unpacker codectypes.AnyUnpacker) err
 func (m MsgSubmitProposal) GetSigners() []sdk.AccAddress {
 	signers := make([]sdk.AccAddress, len(m.Proposers))
 	for i, p := range m.Proposers {
-		proposer, _ := sdk.AccAddressFromBech32(p)
+		proposer := sdk.MustAccAddressFromBech32(p)
 		signers[i] = proposer
 	}
 	return signers
@@ -202,7 +202,7 @@ func (m MsgWithdrawProposal) ValidateBasic() error {
 
 // GetSigners implements Msg.
 func (m MsgWithdrawProposal) GetSigners() []sdk.AccAddress {
-	signer, _ := sdk.AccAddressFromBech32(m.Address)
+	signer := sdk.MustAccAddressFromBech32(m.Address)
 	return []sdk.AccAddress{signer}
 }
 
@@ -231,7 +231,7 @@ func (m MsgVote) ValidateBasic() error {
 
 // GetSigners implements Msg.
 func (m MsgVote) GetSigners() []sdk.AccAddress {
-	signer, _ := sdk.AccAddressFromBech32(m.Voter)
+	signer := sdk.MustAccAddressFromBech32(m.Voter)
 	return []sdk.AccAddress{signer}
 }
 
@@ -252,7 +252,7 @@ func (m MsgExec) ValidateBasic() error {
 
 // GetSigners implements Msg.
 func (m MsgExec) GetSigners() []sdk.AccAddress {
-	signer, _ := sdk.AccAddressFromBech32(m.Signer)
+	signer := sdk.MustAccAddressFromBech32(m.Signer)
 	return []sdk.AccAddress{signer}
 }
 
@@ -269,7 +269,7 @@ func (m MsgLeaveFoundation) ValidateBasic() error {
 
 // GetSigners implements Msg.
 func (m MsgLeaveFoundation) GetSigners() []sdk.AccAddress {
-	signer, _ := sdk.AccAddressFromBech32(m.Address)
+	signer := sdk.MustAccAddressFromBech32(m.Address)
 	return []sdk.AccAddress{signer}
 }
 
@@ -325,7 +325,7 @@ func (m MsgGrant) UnpackInterfaces(unpacker codectypes.AnyUnpacker) error {
 
 // GetSigners implements Msg.
 func (m MsgGrant) GetSigners() []sdk.AccAddress {
-	signer, _ := sdk.AccAddressFromBech32(m.Operator)
+	signer := sdk.MustAccAddressFromBech32(m.Operator)
 	return []sdk.AccAddress{signer}
 }
 
@@ -350,6 +350,6 @@ func (m MsgRevoke) ValidateBasic() error {
 
 // GetSigners implements Msg.
 func (m MsgRevoke) GetSigners() []sdk.AccAddress {
-	signer, _ := sdk.AccAddressFromBech32(m.Operator)
+	signer := sdk.MustAccAddressFromBech32(m.Operator)
 	return []sdk.AccAddress{signer}
 }